server/clickbait_api/apps/score/serializers.py
from rest_framework import serializers
from apps.score.models import Article

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = '__all__'