server/clickbait_api/requirements.txt
tensorflow==1.7.0
Django==2.0.4
h5py==2.7.1
beautifulsoup4==4.6.0
bs4==0.0.1
newspaper3k==0.2.6
PyJWT==1.3.0
konlpy==0.4.4
Keras==2.1.5
JPype1-py3==0.5.5.2
django-cors-headers==2.2.0
djangorestframework==3.8.2
gensim==3.4.0
